requirements.txt
psycopg2-binary==2.8.3
Flask==1.1.1
boto3==1.7.19
botocore==1.10.84